requirements.txt
SQLAlchemy==2.0.23
aiogram==3.1.1
loguru==0.7.2
pydantic==2.3
dependency-injector==4.41.0
python-dotenv~=1.0.0
asyncpg==0.29.0
greenlet==3.0.1
alembic==1.12.1